+++ Unit05-Mechanics-of-trees/homework5b-bst/BST.h
@@ -1,7 +1,5 @@
 #pragma once
 
-#include <cstddef>
-
 template<class T>
 class BST {
 public:
@@ -46,7 +44,7 @@ public:
         // remove all nodes from the tree
     }
 
-    size_t size() const {
+    int size() const {
         // implement size here
         // return the number of nodes in the tree
     }
